Bug:  PRD Reviewer:  Self Comments:  Changes for GWT based binder views.  {95th} 1. Further refinements to the ColumnWidth object, Spinner's and the    column sizing dialog.
Files:
M    gwt/src/org/kablink/teaming/gwt/client/datatable/SizeColumnsDlg.java
M    gwt/src/org/kablink/teaming/gwt/client/widgets/SpinnerListener.java
M    gwt/src/org/kablink/teaming/gwt/client/widgets/ValueSpinner.java
M    gwt/src/org/kablink/teaming/gwt/client/widgets/Spinner.java
M    gwt/src/org/kablink/teaming/gwt/client/widgets/TimePicker.java
M    gwt/src/org/kablink/teaming/gwt/client/binderviews/DataTableFolderViewBase.java
M    gwt/src/org/kablink/teaming/gwt/client/binderviews/folderdata/ColumnWidth.java

+++ trunk/ssf/gwt/src/org/kablink/teaming/gwt/client/widgets/SpinnerListener.java
@@ -62,5 +62,5 @@ public interface SpinnerListener {
 	 * 
 	 * @param value	The current spinner value.
 	 */
-	void onSpinning(long value);
+	void onSpinning(double value);
 }
